Restructured tests to use the new runalltests() function
